tesla-parts-backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
requests
python-dotenv
pytest
httpx
